Clean up authenticator
